Add functions to sort array using quicksort.
Sorting involves three functions:
- quick_sort(): check size of array. If small, defer to insertion sort.
  If large enough, begin recursive quicksort process by calling
  quick_sort_sort().
- quick_sort_sort(): performs quicksort recursively.
- quick_sort_partition(): partitions subarray around pivot and
  returns index of pivot.
Having this separation gives greater control over the sort and also
keeps the quick_sort() function header the same as the other sorts.

In addition to these functions I've also added the helper function
median_three(). The function finds the median of three elements in an
array and returns the index of the median element. When finding pivot
points for subarrays in quicksort, it turns out that choosing the
median of the indices (lo, mid, hi) gives a better pivot than
always picking any one of these values.