Add linked-list stack implementation
Why the change is needed:

Timsort requires a stack onto which it can push / pop runs in order to
perform its merges.

How the change will help:

Current stack implementation should be sufficient for Timsort. Stack
assumes that data values are stored in memory stack which, given that
runs will be represented by small TimsortRun structs, shouldn't be a
problem.

Extra:

I've also reorganized the project's directory structure. The spec
directory has been moved to the project's root directory and uses a
Makefile similar to the one in the root directory to build a test
executable.