Add outline of functions to perform Timsort
Timsort is perfomred using the following 4 functions:
- timsort()
- timsort_minrun()
- timsort_find_runs()
- timsort_merge_runs()
In addition, I've decided to limit TimsortRun structs to only the
members 'start' and 'len'. These two values should be sufficient to
identify and delimit any run.

It is highly unlikely that these functions will remain exactly as
written, but their structure / roles more-or-less should.