Fix Timsort implementation
Why the change is needed:

The current implenetation of Timsort is not quite correct - while its
general structure is fine, it doesn't actually work.

How the change will help:

The functions comprsing Timsort have been fixed and are now producing
correct results. In particular, I've found and addressed two major
errors with the previous implementation:
- Runs were not being properly handled. In particular, nothing was done
to account for the final run (which could be less than minrun in
length), and descending runs were not being reversed.
- Merges were not being performed correctly. In particular, I was using
the wrong set of indicies with the temporary copy of the smaller run (I
was forgetting that temporary run had initial index 0 while the smaller
run could have some other index).

While Timsort is now working, it is nowhere near as efficient as it
should be.