Add helper function for performing binary search
Why the change is needed:

The merging process can be improved by using binary search.
Specifically, for two sorted runs A and B, we use binary search to find
the indices of the first and last elements of A in B. We can then ignore
any values outside of this range.