Improve Timsort function for merging runs
Why the change is needed:

The current implementation of timsort_merge_runs() is not as efficient
as it should be. In particular, it merges more elements than it needs
to, requiring both more time and more memory.

How the change will help:

When merging two consecutive runs A and B, as A and B are sorted, we
do not need to consider values in A which are smaller than the smallest
element of B, nor elements in B larger than the largest element of A.
The upated timsort_merge_runs() function uses binary search to find
the indices of these values and then only merges elements which are
between the indices. While this has little effect on random data, it
makes merging much more efficient for natural data.

Extra:

I've also made a correction to bin_search(). I'd previously made the
values 'lo' and 'hi' of type size_t. This was creating problems,
as 'hi' was occassionally decremented when it was 0 causing it to wrap
to a value outside the bounds of the array. To fix this, I've made both
'lo' and 'hi' integers and have added a condition to check if either are
negative.