Fix use of binary search in Timsort
Why the change is needed:

Timsort uses binary search in two ways:
- Indirectly during binary insertion sort when sorting runs.
- Directly while merging runs to find lower and upper index bounds
  in which to merge.
However, binary search does not function the same way in both cases.
In the former, it doesn't expect the element to be in the array and
instead returns where it WOULD be located in the array. In the second
case, it does expect the element to be in the array and returns where
the element IS located in the array (if element isn't found, returns
-1).
Currently, binary search only functions in the first manner.

How the change will help:

The binary_insert_sort() function now implements its own version of
binary search to satisfy its particular requirements. The bin_search()
function offers the more typical implentation of binary search and is
used during timsort_merge_runs().