Improve Timsort function for finding runs
Why the change is needed:

The current timsort_find_runs() function is incredibly bloated, mostly
from the forloop used to track runs.

How the change will help:

I've significantly condensed the forloop by recognizing and removing
duplicate code.

The previous forloop contained a conditional that checked:
- If run is currently descending.
- Else run is currently non-descending.

Inside each condition a second conditional checked:
- If the run continues (curr element shares same relation to previous
  element as next element does to curr element).
- Else the curr element terminates the run.

I've realized that first conditional is unnecessary, as we only need to
track whether or not the run continues. The exact orientation of the run
only matters in whether or not it is reversed - all other functionality
- incrementing run len while run continues, adding current run to stack
when run terminates, etc. - is the same for both.

Extra:

In addition to the above change, I've also made the following small
improvements:
- TimsortRuns are now initialized using an initializer list rather than
  setting members individually. This was done mostly for aesthetic
  reasons.
- I've added a 'check_runs' value to limit the amount that the
  invariants are checked. The invariants are now checked only after a
  run has been added to the runs stack, rather than after each
  iteration of the forloop.