Split up Timsort merge runs function
Why the change is needed:

The current timsort_merge_runs() function is already rather large. As I
am planning on implementing galloping soon, which will make the function
even longer, I believe it makes sense to reorganize and condense the
function.

How the change will help:

The timsort_merge_runs() function now makes use of helpers
timsort_merge_runs_lo() and timsort_merge_runs_hi(). The former
merges left to right and is used if the first run is the shorter run,
while the latter merges right to left and is used if the second run is
shorter.

Extra:

I've also realized that I made a mistake in timsort_find_runs(). When
checking to see if either of the invariants |X| > |Y| + |Z| or |Y| > |Z|
had failed, I should have used '<=' instead of just '<'.