Optimize how Timsort merges runs
Why the change is needed:

When merging two runs, it is occasionally the case that one run is both
longer than, and contains many elements outside the bounds of,
the other run. When this happens, the smaller array is used up early in
the merge and the rest of the merge is spent iterating through elements
of the larger array.

How the change will help:

When a case like this arises, the updated timsort_merge_runs_[hi, lo]()
functions will now perform a bulk merge of the remaining elements in the
larger array. This provides a noticeable (~10%) performance boost.