Add stack function to pop and return
Why the change is needed:

This function isn't really needed, I just prefer getting values from a
stack using a single function rather than having to combine peek() and
pop() (although the latter method is safer).

Extra:

I've updated timsort_find_runs() to use this new function. In addition,
I've slightly tweaked some of the other stack functions. In particular,
I've changed the 'free' functions so that they now take a double pointer
rather than a pointer. This allows the 'free' functions to both free the
memory referenced by a pointer and set the pointer to null. Not entirely
sure if I'll stick with this schema in the long run but I'm going to try
it for now.