Restructure Timsort funcs to prepare for galloping
Why the change is needed:

To implement galloping mode, certain information needs to persist
between merges (e.g. is merging currently in galloping mode?). Other
functions (e.g. timsort_merge_runs) will need to be altered to accept
this information as a paramter.

How the change will help:

I've created a struct 'TimsortMergeState' to contain this information -
currently it includes the run stack and the current value of
'min_gallop'. As mentioned, merge state should persist for the entirety
of the merge. To make this clear, TimsortMergeState is initialized
during the call to timsort(), and is then passed to functions which need
it.

Extra:

In addition, I've taken the opportunity to split up timsort_find_runs().
I felt that the function had too many responsibilities:
- it was finding runs
- it was checking run invariants
- it was merging all runs together
As such, I've made two new functions, timsort_check_invariants() and
timsort_collapse_runs(), to handle these other two responsibilities.