Expand Timsort merge state struct
Why the change is needed:

The current merge state struct has two members:
- Runs stack
- Minimum gallop threshold

With respect to the runs stack, I've realized that I made a huge
oversight when I separted timsort_collapse_runs() into its own
function. Currently, runs are local to timsort_find_runs().
Consequently, their memory is freed once the function has finished
executing. This wasn't a problem before, as checking invariants /
collapsing runs was performed inside of timsort_find_runs() (which is
why timsort_check_invariants() still works). However, this means that
timsort_collapse_runs() does not have access to the runs - inside of the
function, runs in the run stack point to freed memory.

In addition, the merge state struct needs a member to indicate if
galloping mode is on or off.

How the change will help:

To address the problem with the runs stack, I've moved initialization of
the run array to timsort() and made the array a member of merge state.
By doing this, runs will persist until the sort completes and
the other Timsort functions will be able to access them.

To track if galloping mode is on or off, I've added a member 'galloping'
to the merge state.