Implement galloping mode during merge lo
Why the change is needed:

Timsort is all about sorting real-world data, but the current merge
functions do not yet reflect this. In particular, it is common in
real-world data for many of the values in one run to be smaller than
another. In a standard merge, we do not account for this - elements are
compared one-to-one, requiring one comparison for every element in the
run.

How the change will help:

Galloping mode allows us to improve on this signficantly (at least for
non-random data). Galloping mode is triggered when the number of
consecutive elements selected from one of the runs exceeds MIN_GALLOP.
In galloping mode, for runs A (left, smaller) and B,
we perform two searches:
- location of A[0] in B
- location of B[0] in A
Once one of the locations has been found, all of the elements of the
array being searched up to that location (slice{1,2} := number of such
elements) are copied into the main array in bulk. The value being
searched for is then copied in and the main indices (current index
in main array for copying, initial indicies for A and B) are
incremented to reflect the number of copied elements.
After this, we check that slice1 and slice2 exceed MIN_GALLOP. If they
didn't, then galloping mode was not efficient and we both increment
MIN_GALLOP (to make entering into galloping mode harder) and exit from
galloping mode. If they do, we decrement MIN_GALLOP to make entering
into galloping mode easier. In this way, the sort can quickly adapt to
data for which galloping mode is either efficient / inefficient.