Change indices in merge hi func from size_t to int
Why the change is needed:

The current timsort_merge_runs_hi() function does not work as intended.
The merge loop contains an if statement which is supposed to check
(among other things) that the rightmost run still contains unmerged
values. It does this by checking that r (the current index of the right
run) is >= 0. However, because r is of type size_t, I had previously
added a condition that r should only be decremented if it is greater
than 0. This was an oversight on my part, as it means r is always >= 0.

How the change will help:

In order to avoid this problem, I've changed the indices used in the
merge loop (k, l and r) to be integers rather than size_t. This way they
can be decremented without issue. Because of this change, I've also
needed to update the part of the condition which checks that l (an
integer) is less than lo (size_t). The condition now checks if either l
< 0, or if not, l < lo.