Consolidate documentation into just source files
Why the change is needed:

Currently, both the source and header files are documented. However,
much of the documentation in the header files is either redundant (as
it duplicates what already exists in the source files) or is out of
date.

How the change will help:

I've removed most of the documentation from the headers (I've left some
for structs and #define directives). Having source files
as the sole source of documentation removes this redundancy and
increases the chance that documentation will be kept up to date.