Refactor insertion sort functions
Why the change is needed:

insert_sort_partial():

The current implementation of insert_sort_partial() uses a clunky
while loop for performing swaps. Because a comparison of pointers 'a'
and 'b' is part of the loop condition, memory addresses need to be
assigned to these pointers both before and inside of the loop. In
addition, the current index variables are integers but really ought to
be size_t, as I've decided to treat indices universally as size_t.

binary_insert_sort():

The current implemenation of binary_insert_sort() contains its own
version of binary_search(). Given how binary search is frequently used
amongst other sorts and hence ought to be its own function, this is
totally unnecessary. In addition, the function shifts elements
one-by-one when making room for an insertion. This is extremely
inefficient and slows down the function.

How the change will help:

insert_sort_partial():

- While loop now only checks if current index is valid - comparison of
  pointers 'a' and 'b' now takes place inside of the loop. By doing this,
  these pointers do not need to be assigned memory addresses before the
  loop starts.
- Change all indices to size_t.
  - Account for this change by changing while loop condition from size_t
    unsafe 'index > lower_bound' suntax to size_t safe
    'index --> lower_bound' syntax.

binary_insert_sort():

- Now defers to a separate binary_search_loc() function for
  perfomring binary search.
- Rather than shift elements one at a time (e.g. array[i+1] = array[i]),
  now uses memmove() to shift elements in bulk.
    - Significantly improves performance of function.
    - Allows 'j' index (previosuly used when shifting elements) to be
      removed entirely.

Extra:

- Added binary_search_loc() function for performing binary searches.
  I've realized that all of the sorting algorithms which use binary
  search use it to find the hypothetical index of some value in an
  array. They're not checking if the array actually contains the value.
  As such, I'll be replacing all other binary search functions
  (bin_search() and hopefully timsort_binary_search()) with this
  function.
- Added tests for insertion sort functions.
- Update spec Makefile to include std=c99 flag.