Add generic tests for sort functions
Why the change is needed:

Previously, each sorting function required its own test function.
However, each test function only differed in two ways:
- sorting function to test
- arguments to call sorting function with
This meant I was more-or-less repeating the same code over and over for
each test.

How the change will help:

For the most part, the test functions can be categorized as either:
- functions which sort whole arrays and hence have no bounds arguments.
- functions which sort subarrays and hence have bounds arguments.
As such, I've created two generic test functions to handle these cases,
test_sort_no_bounds() and test_sort_bounds().
In order for these functions to work, they need to be passed a function
pointer to a sort function. To be able to do this during a test, I've
added a new macro "mu_run_test_on_arg" to minunit.h. The macro accepts
two arguments, which when used with the above functions correspond to
a pointer to a sort function and that function's name.