Refactor comb sort function
Why the change is needed:

The current comb_sort() implementation has a few problems.
As with selec_sort() before it (and most other sort functions currently),
comb_sort() uses an inefficient process for generating memory offsets.
Specifically, it uses a loop with an integer index and multiples this
value by the size of an element to create an offset (e.g.
pointer+(int_index * element_size)).
More importantly, the main index 'i' is not incremented correctly during
the while-loop. Currently, 'i' is incremented by 'gap' when arr[i] <
arr[i + gap]. This should not happen, as by skipping the values between
'i' and 'i + gap' the loop requires significantly more iterations to
sort the array.

How the change will help:

- The function now treats indices and related values as memory offsets.
  For example, the main index 'i' is now incremented by the 'size'
  parameter (the size of an element in the array) after each iteration
  of the loop, rather lthan by 1. By doing this, memory offsets can be
  found with 1 operation rather than 2 as before.
- The 'i' index is no longer incremented when arr[i] < arr[i + gap].