Add conditional insert sort back to merge sort
Why the change is needed:

To optimize merge_sort_recursive(), the function should defer to
insert_sort() when the current subarray is small. However, I'd commented
out this functionality in a recent commit because I was making changes
to merge_sort_recursive() that where not compatible with insert_sort().
The functions are now compatible and so the code can be uncommented.